fix: rename repo-base util
