+++ .tokeignore
@@ -1,1 +1,3 @@
 yarn.lock
+*.md
+*.mdx
