+++ libs/repo-base/src/index.ts
@@ -1,4 +1,4 @@
 export { useDeleteAll } from './delete-all'
-export { getDataSource } from './get-source'
+export { getDataSource } from './get-data-source'
 export { getRelationNames } from './get-relation-names'
 export { useFindOneByKey } from './find-by-key'
